refactor: 切换数据库从PostgreSQL到MySQL并重构相关代码
主要变更：
1. 替换依赖库，从psycopg2/sshtunnel改为PyMySQL+DBUtils
2. 新增docker-compose.yml配置本地MySQL 8.0开发环境
3. 更新配置文件和环境变量，适配MySQL连接参数
4. 重构数据库操作代码，移除SSH隧道相关逻辑
5. 修正SQL语法：替换ILIKE为LIKE、移除PostgreSQL特有语法
6. 更新API接口路径，调整前端请求地址
7. 重构数据库初始化schema，适配MySQL语法和约束
8. 统一数据库返回值处理逻辑

+++ backend/app_core/services/teacher_service.py
@@ -132,13 +132,13 @@ class TeacherService:
                 c.name,
                 c.course_code,
                 COUNT(e.id) AS enrolled_count,
-                ROUND(AVG(COALESCE(e.final_grade, e.grade))::numeric, 2) AS avg_grade,
+                ROUND(AVG(COALESCE(e.final_grade, e.grade)), 2) AS avg_grade,
                 ROUND(
-                    (SUM(CASE WHEN COALESCE(e.final_grade, e.grade) >= 60 THEN 1 ELSE 0 END)::numeric
+                    (SUM(CASE WHEN COALESCE(e.final_grade, e.grade) >= 60 THEN 1 ELSE 0 END)
                         / NULLIF(COUNT(e.id), 0)) * 100, 2
                 ) AS pass_rate,
                 ROUND(
-                    (SUM(CASE WHEN COALESCE(e.final_grade, e.grade) >= 90 THEN 1 ELSE 0 END)::numeric
+                    (SUM(CASE WHEN COALESCE(e.final_grade, e.grade) >= 90 THEN 1 ELSE 0 END)
                         / NULLIF(COUNT(e.id), 0)) * 100, 2
                 ) AS excellent_rate
             FROM courses c
